+++ api-tests/relationships/filtering/filtering.test.js
@@ -1,5 +1,6 @@
 const { Text, Relationship } = require('@keystonejs/fields');
-const { multiAdapterRunners, setupServer, graphqlRequest } = require('@keystonejs/test-utils');
+const { multiAdapterRunners, setupServer } = require('@keystonejs/test-utils');
+const { createItem } = require('@keystonejs/server-side-graphql-client');
 
 function setupKeystone(adapterName) {
   return setupServer({
@@ -31,15 +32,30 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
     describe('relationship filtering', () => {
       test(
         'nested to-single relationships can be filtered within AND clause',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          const company = await create('Company', { name: 'Thinkmill' });
-          const otherCompany = await create('Company', { name: 'Cete' });
-
-          const user = await create('User', { company: company.id });
-          await create('User', { company: otherCompany.id });
-
-          const { data, errors } = await graphqlRequest({
+        runner(setupKeystone, async ({ keystone }) => {
+          const company = await createItem({
             keystone,
+            listKey: 'Company',
+            item: { name: 'Thinkmill' },
+          });
+          const otherCompany = await createItem({
+            keystone,
+            listKey: 'Company',
+            item: { name: 'Cete' },
+          });
+
+          const user = await createItem({
+            keystone,
+            listKey: 'User',
+            item: { company: { connect: { id: company.id } } },
+          });
+          await createItem({
+            keystone,
+            listKey: 'User',
+            item: { company: { connect: { id: otherCompany.id } } },
+          });
+
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers(where: {
@@ -76,15 +92,30 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
 
       test(
         'nested to-single relationships can be filtered within OR clause',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          const company = await create('Company', { name: 'Thinkmill' });
-          const otherCompany = await create('Company', { name: 'Cete' });
-
-          const user = await create('User', { company: company.id });
-          await create('User', { company: otherCompany.id });
-
-          const { data, errors } = await graphqlRequest({
+        runner(setupKeystone, async ({ keystone }) => {
+          const company = await createItem({
             keystone,
+            listKey: 'Company',
+            item: { name: 'Thinkmill' },
+          });
+          const otherCompany = await createItem({
+            keystone,
+            listKey: 'Company',
+            item: { name: 'Cete' },
+          });
+
+          const user = await createItem({
+            keystone,
+            listKey: 'User',
+            item: { company: { connect: { id: company.id } } },
+          });
+          await createItem({
+            keystone,
+            listKey: 'User',
+            item: { company: { connect: { id: otherCompany.id } } },
+          });
+
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
           query {
             allUsers(where: {
@@ -121,20 +152,29 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
 
       test(
         'nested to-many relationships can be filtered within AND clause',
-        runner(setupKeystone, async ({ keystone, create }) => {
+        runner(setupKeystone, async ({ keystone }) => {
           const ids = [];
 
-          ids.push((await create('Post', { content: 'Hello world' })).id);
-          ids.push((await create('Post', { content: 'hi world' })).id);
-          ids.push((await create('Post', { content: 'Hello? Or hi?' })).id);
+          ids.push(
+            (await createItem({ keystone, listKey: 'Post', item: { content: 'Hello world' } })).id
+          );
+          ids.push(
+            (await createItem({ keystone, listKey: 'Post', item: { content: 'hi world' } })).id
+          );
+          ids.push(
+            (await createItem({ keystone, listKey: 'Post', item: { content: 'Hello? Or hi?' } })).id
+          );
 
-          const user = await create('User', { posts: ids });
+          const user = await createItem({
+            keystone,
+            listKey: 'User',
+            item: { posts: { connect: ids.map(id => ({ id })) } },
+          });
 
           // Create a dummy user to make sure we're actually filtering it out
-          await create('User', { posts: [] });
+          await createItem({ keystone, listKey: 'User', item: {} });
 
-          const { data, errors } = await graphqlRequest({
-            keystone,
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers(where: {
@@ -155,34 +195,38 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
 
           expect(errors).toBe(undefined);
           expect(data).toHaveProperty('allUsers.0.posts');
+          expect(data.allUsers).toHaveLength(1);
+          expect(data.allUsers[0].id).toEqual(user.id);
           expect(data.allUsers[0].posts).toHaveLength(3);
-          expect(data).toMatchObject({
-            allUsers: [
-              {
-                id: user.id,
-                posts: [{ id: ids[0] }, { id: ids[1] }, { id: ids[2] }],
-              },
-            ],
-          });
+          expect(data.allUsers[0].posts.map(({ id }) => id).sort()).toEqual(ids.sort());
         })
       );
 
       test(
         'nested to-many relationships can be filtered within OR clause',
-        runner(setupKeystone, async ({ keystone, create }) => {
+        runner(setupKeystone, async ({ keystone }) => {
           const ids = [];
 
-          ids.push((await create('Post', { content: 'Hello world' })).id);
-          ids.push((await create('Post', { content: 'hi world' })).id);
-          ids.push((await create('Post', { content: 'Hello? Or hi?' })).id);
+          ids.push(
+            (await createItem({ keystone, listKey: 'Post', item: { content: 'Hello world' } })).id
+          );
+          ids.push(
+            (await createItem({ keystone, listKey: 'Post', item: { content: 'hi world' } })).id
+          );
+          ids.push(
+            (await createItem({ keystone, listKey: 'Post', item: { content: 'Hello? Or hi?' } })).id
+          );
 
-          const user = await create('User', { posts: ids });
+          const user = await createItem({
+            keystone,
+            listKey: 'User',
+            item: { posts: { connect: ids.map(id => ({ id })) } },
+          });
 
           // Create a dummy user to make sure we're actually filtering it out
-          await create('User', { posts: [] });
+          await createItem({ keystone, listKey: 'User', item: {} });
 
-          const { data, errors } = await graphqlRequest({
-            keystone,
+          const { data, errors } = await keystone.executeGraphQL({
             query: `
           query {
             allUsers(where: {
@@ -203,49 +247,87 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
 
           expect(errors).toBe(undefined);
           expect(data).toHaveProperty('allUsers.0.posts');
+          expect(data.allUsers).toHaveLength(1);
+          expect(data.allUsers[0].id).toEqual(user.id);
           expect(data.allUsers[0].posts).toHaveLength(3);
-          expect(data).toMatchObject({
-            allUsers: [
-              {
-                id: user.id,
-                posts: [{ id: ids[0] }, { id: ids[1] }, { id: ids[2] }],
-              },
-            ],
-          });
+          expect(data.allUsers[0].posts.map(({ id }) => id).sort()).toEqual(ids.sort());
         })
       );
 
       test(
         'many-to-many filtering composes with one-to-many filtering',
-        runner(setupKeystone, async ({ keystone, create }) => {
-          const adsCompany = await create('Company', { name: 'AdsAdsAds' });
-          const otherCompany = await create('Company', { name: 'Thinkmill' });
+        runner(setupKeystone, async ({ keystone }) => {
+          const adsCompany = await createItem({
+            keystone,
+            listKey: 'Company',
+            item: { name: 'AdsAdsAds' },
+            returnFields: 'id name',
+          });
+          const otherCompany = await createItem({
+            keystone,
+            listKey: 'Company',
+            item: { name: 'Thinkmill' },
+            returnFields: 'id name',
+          });
 
           // Content can have multiple authors
-          const spam1 = await create('Post', { content: 'spam' });
-          const spam2 = await create('Post', { content: 'spam' });
-          const content = await create('Post', { content: 'cute cat pics' });
+          const spam1 = await createItem({ keystone, listKey: 'Post', item: { content: 'spam' } });
+          const spam2 = await createItem({ keystone, listKey: 'Post', item: { content: 'spam' } });
+          const content = await createItem({
+            keystone,
+            listKey: 'Post',
+            item: { content: 'cute cat pics' },
+          });
 
-          const spammyUser = await create('User', {
-            company: adsCompany.id,
-            posts: [spam1.id, spam2.id],
+          const spammyUser = await createItem({
+            keystone,
+            listKey: 'User',
+            item: {
+              company: { connect: { id: adsCompany.id } },
+              posts: { connect: [{ id: spam1.id }, { id: spam2.id }] },
+            },
           });
-          const mixedUser = await create('User', {
-            company: adsCompany.id,
-            posts: [spam1.id, content.id],
+          const mixedUser = await createItem({
+            keystone,
+            listKey: 'User',
+            item: {
+              company: { connect: { id: adsCompany.id } },
+              posts: { connect: [{ id: spam1.id }, { id: content.id }] },
+            },
           });
-          const nonSpammyUser = await create('User', {
-            company: adsCompany.id,
-            posts: [content.id],
+          const nonSpammyUser = await createItem({
+            keystone,
+            listKey: 'User',
+            item: {
+              company: { connect: { id: adsCompany.id } },
+              posts: { connect: [{ id: content.id }] },
+            },
           });
-          const quietUser = await create('User', { company: adsCompany.id, posts: [] });
-          await create('User', { company: otherCompany.id, posts: [content.id] });
-          await create('User', { company: otherCompany.id, posts: [spam1.id] });
+          const quietUser = await createItem({
+            keystone,
+            listKey: 'User',
+            item: { company: { connect: { id: adsCompany.id } } },
+          });
+          await createItem({
+            keystone,
+            listKey: 'User',
+            item: {
+              company: { connect: { id: otherCompany.id } },
+              posts: { connect: [{ id: content.id }] },
+            },
+          });
+          await createItem({
+            keystone,
+            listKey: 'User',
+            item: {
+              company: { connect: { id: otherCompany.id } },
+              posts: { connect: [{ id: spam1.id }] },
+            },
+          });
 
           // adsCompany users whose every post is spam
           // NB: this includes users who have no posts at all
-          let { data, errors } = await graphqlRequest({
-            keystone,
+          let { data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers(where: {
@@ -275,8 +357,7 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
           ]);
 
           // adsCompany users with no spam
-          ({ data, errors } = await graphqlRequest({
-            keystone,
+          ({ data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers(where: {
@@ -308,8 +389,7 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
           ]);
 
           // adsCompany users with some spam
-          ({ data, errors } = await graphqlRequest({
-            keystone,
+          ({ data, errors } = await keystone.executeGraphQL({
             query: `
         query {
           allUsers(where: {