+++ packages/fields-oembed/src/Implementation.js
@@ -29,6 +29,10 @@ export class OEmbed extends Implementation {
     this.parameters = parameters;
   }
 
+  get _supportsUnique() {
+    return false;
+  }
+
   gqlOutputFields() {
     return [`${this.path}: ${this.graphQLOutputType}`];
   }
@@ -285,7 +289,7 @@ export class KnexOEmbedInterface extends CommonOEmbedInterface(KnexFieldAdapter)
 
     // Error rather than ignoring invalid config
     // We totally can index these values, it's just not trivial. See issue #1297
-    if (this.config.isUnique || this.config.isIndexed) {
+    if (this.config.isIndexed) {
       throw `The OEmbed field type doesn't support indexes on Knex. ` +
         `Check the config for ${this.path} on the ${this.field.listKey} list`;
     }