+++ packages/fields/src/types/Decimal/test-fixtures.js
@@ -5,6 +5,8 @@ import Decimal from './';
 export const name = 'Decimal';
 export { Decimal as type };
 export const exampleValue = '"6.28"';
+export const exampleValue2 = '"6.283"';
+export const supportsUnique = true;
 
 export const getTestFields = () => {
   return {
@@ -116,8 +118,8 @@ export const filterTests = withKeystone => {
 
 export const crudTests = withKeystone => {
   const withHelpers = wrappedFn => {
-    return async ({ keystone, listName, ...rest }) => {
-      const list = keystone.getListByKey(listName);
+    return async ({ keystone, listKey, ...rest }) => {
+      const list = keystone.getListByKey(listKey);
       const { listQueryName } = list.gqlNames;
       const {
         data: { [listQueryName]: items },
@@ -132,7 +134,7 @@ export const crudTests = withKeystone => {
         }`,
       });
       expect(errors).toBe(undefined);
-      return wrappedFn({ keystone, listName, ...rest, list, items, ...list.gqlNames });
+      return wrappedFn({ keystone, listKey, ...rest, list, items, ...list.gqlNames });
     };
   };