+++ packages/fields/src/types/Float/Implementation.js
@@ -8,6 +8,10 @@ export class Float extends Implementation {
     this.isOrderable = true;
   }
 
+  get _supportsUnique() {
+    return true;
+  }
+
   gqlOutputFields() {
     return [`${this.path}: Float`];
   }