+++ packages/fields/src/types/Url/test-fixtures.js
@@ -1,0 +1,7 @@
+import Url from './';
+
+export const name = 'Url';
+export { Url as type };
+export const exampleValue = '"https://keystonejs.org"';
+export const exampleValue2 = '"https://thinkmill.com.au"';
+export const supportsUnique = true;