+++ packages/fields/src/types/Uuid/Implementation.js
@@ -15,6 +15,10 @@ export class UuidImplementation extends Implementation {
     this.isOrderable = true;
   }
 
+  get _supportsUnique() {
+    return true;
+  }
+
   gqlOutputFields() {
     return [`${this.path}: ID`];
   }